tests: move TestBlocks to testing::sync
