+++ testing/src/network/mod.rs
@@ -14,9 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod blocks;
-pub use blocks::*;
-
 #[cfg(test)]
 pub mod encryption;
