+++ testing/src/network/sync.rs
@@ -181,7 +181,7 @@ async fn block_two_node() {
 
     const NUM_BLOCKS: usize = 100;
 
-    let blocks = crate::network::TestBlocks::load(NUM_BLOCKS).0;
+    let blocks = crate::sync::TestBlocks::load(NUM_BLOCKS).0;
     assert_eq!(blocks.len(), NUM_BLOCKS);
 
     for block in blocks {
